added rock components, cleaned up, move data to separate file
