+++ lib/giveData.js
@@ -1,0 +1,81 @@
+/**
+ * note : This is all the hardcoded data for the Give page.
+ */
+
+const otherWaysToGiveData = [
+  {
+    icon: 'envelope-open-dollar',
+    title: 'Give Cash or Check',
+    htmlContent: `<div style="color: black; font-style: italic;">Give at any <a href="/locations" color: #004F71;">Christ Fellowship location</a> or mail to: <br/><br/>
+      Christ Fellowship Church Contributions <br/>
+      5343 Northlake Blvd.Palm Beach Gardens, FL 33418</div>`,
+  },
+  {
+    icon: 'plus',
+    title: 'Stocks, Bonds, Crypto & other Assets',
+    htmlContent: `<div style="color: black; font-style: italic;">To give by stock, bond, or crypto currency please <a href="#contact-us??" style="color: #004F71;">contact us.</a></div>`,
+  },
+];
+const tithingOfferingCards = [
+  {
+    title: 'Tithing',
+    subtitle: 'TRUSTING GOD WITH YOUR FINANCES.',
+    htmlContent: `Tithing is a biblical principle that means the tenth. The first tenth, which belongs to God, is Holy and set apart for Him. When we bring our tithe to God, it means that we trust and acknowledge that He’s the giver of every good thing in our lives. Tithing isn’t as much about finances as it is about faith. It’s not about what God wants from you but what He has for you.<br/><br/> In <span style="font-weight: bold;">Malachi 3:10</span>, the Bible says we can test this promise. When we bring our first and best back to God, He promises to bless the rest of our resources so that we could be a blessing toward others.<br/><br/> Is tithing new for you? Take the <a href="https://rock.gocf.org/310challenge">Malachi 3:10 Challenge</a> today!`,
+  },
+  {
+    title: 'Offerings',
+    subtitle: 'GIVING BEYOND THE TITHE.',
+    htmlContent: `As Pastor Todd has shared, "We're never more like Jesus than when we serve and give." In scripture, we're told that God so loved the world that He gave. Our offerings go beyond the obedience of the tithe; they are a reflection of a life marked by generosity. We believe generosity has a divine purpose connected to it, which is why we invite our church family to give beyond the tithe in one of these ways throughout the year. <br/><br/>Learn more about two special offerings we do each year: <a href="#???">Heart for the House and Christ Birthday Offering.</a>`,
+  },
+];
+
+const giveImpactData = [
+  {
+    icon: 'campus',
+    title: 'Church Operations',
+    htmlContent: `<div style="color: black; margin-bottom: 20px; font-weight: bold;">From maintaining our facilities, supporting our staff, and carrying out various ministries effectively, we’re able to bring the love and message of Jesus to our region.</div>`,
+  },
+  {
+    icon: 'users',
+    title: 'Weekly Ministry Offerings',
+    htmlContent: `<div style="color: black; margin-bottom: 20px; font-weight: bold;">We have ministry offerings for the entire family all throughout the week to help people grow in community and in their relationship with God.</div>`,
+  },
+  {
+    icon: 'handshake',
+    title: 'Local and Global Missions',
+    htmlContent: `<div style="color: black; margin-bottom: 20px; font-weight: bold;">Through service projects, trips, and supporting our partners, we’re able to serve others at their point of need and share the love of Jesus in tangible ways.</div>`,
+  },
+  {
+    icon: 'computer',
+    title: 'Digital Resources',
+    htmlContent: `<div style="color: black; font-weight: bold;">With weekly live online services and other digital resources like online groups and classes, and free articles and devotionals, people can grow in their faith from anywhere. </div>`,
+  },
+];
+
+const carouselCards = [
+  {
+    wistiaId: 'kpga7mn5cj',
+    image:
+      'https://embed-ssl.wistia.com/deliveries/8a1ca2f5766877dc513d1b80b373d0e3.jpg',
+  },
+  {
+    wistiaId: 'dbukx4kdmk',
+    image:
+      'https://embed-ssl.wistia.com/deliveries/f82aa181462b90aeebb32438efd8d0ee.jpg',
+  },
+];
+
+const moreGivingOpportunitiesId =
+  'HorizontalCardListFeature:f6b30a38eb547f1f75676ce0557a7e05bed66de665564bd0c5e37f0692676795e12dca024e16de4c6d7782ed0288eb10570bc45a488bee66c93ee7b8735beb76978e3501e90a70caa695d63b60072fa825dc6307ecc9345b7ff0ca86896f2b1362a2bb6e9837de1c1384faa2021dbc48e36ee7d92b7b3466b8bfea2f9e6338890622ddfb473bd8fcac042699abfc0b26645562e66bf13a58993422bddef2f2608e74a71de67e33359e921316a76e60aac615240bc48f207d1d11ac5743e97ac8d33ac1781789821227c075123e0111a2902a8579297e49d83d6f49894dbea5aa58e52fbf33f764900f9a17f899d20de8f25ed40ec706c9046dbebfae1649934d6df79a81de611d153a1b8f80376636e4055901c9b1008c28473d32b6b3f0999476dd1d4c38bd9cbc64e91e377b2d69fa778032614265b750ab47213a2234a8a25f40b2d1501edc117e150a1c2f89941dd9e70db6d9219ec5f286deedadc4dac5991617e132de6df6e495fa578c7402cda7dae3a4ceefc22f6656cb39f2093e1efd7566b73382f2bed22176da77ae3574edc51b572008a14651c341a876ea2f94dd2373fa1ba125ff7f6cd9ce674041425e9f6969a5cba9781961a9202a2a1acefa3ec1340322f5a11a025116b3ce5fea1861885af2936b3768d93e39e48aaa93ad8054b3cdf37dbe3bb3044c512bd8262d5a0ee1cd25e3f13e5d7f22161625bbed46a016afa0873bb216a9d416761e8a7b9b2d121b2da4e3d3099f833e0edb4727ffed670a236db6ef512e1cb526652ffbf0f2dcf66d164d6c9ce57bd7bc150b789e33e8861124d744bf486af7c8982a3579d1f392428e1c58d5a8afcee2fd2c6fa251f03980dac93f0348dd3530dbfc42e58a3da21e55577057c3ae0a90716bc62390d635634a0b7f33793c67769166c92517ec0960382f4d393c9035bed8d292b047981f0191b4ddf28e0b62a1806387da969793c069e7dc6857ed2a632b73f197530b28737eed4cf10f45cb527ad38a796ff10d12b67f570ca363f7916ba31a3810f74a0de6a590fc23f7b97f6b583751b1281049bf3391f78d54246023c9a817c51aaf635f2d9fed30aa94cc55532e0c33546208d66a676ffb02faa1597711a476294e87762e68f632540e3beb1544091afd4e4281db94e0f3b5f6e489c9fa598a48e8caa1941c8bc5a97646593b876d9694711f3d2128cb5517ff8eaf0a657cc0ba03e3268b169917c88c67141485a8b55195becf97475c72f6dea17533a804f91fc677b4ebb003f1db0e62c436a0159da684edff3fa2331d6de331be06232a54ca7b5fec0bfe09f9b8c08a428b6542728a44ccd3914cb4ccb5c1e4f2550e748608be70cadb3247930a72a07d3585e93f5ab0e0184558cb4ab1b313d0ed3073c6a1fb50602ea83e461a795ceb19cbad4aa8e8f2d82bf79522bdc9004d869d9da3103b85d41d6590b2ea7c8f19e6e2fb9e45b886120c5dd5399d660accc9b9e0b06b58f7eca93485998b4f903748c6091c642452e8ff5e2c50738ccc7776b8ac065ecff76d726ebf45be85e8f25e3621a2cb269fad3fbded3b4cb911cea9e46c68c3adc0ca192b290c01d2005ecf4c0670bf3489a85b6ee45d577c32d9ce6766f47edd99e3661306bbc34cd65a884c066e6b39e31fddd6c94654a690378447f89bb65c63762f7c829934dbfcb00fde1ee468004a816522c3fa3891a9cec286c4a1b6daab6ed571b12389fabd9eaa330729f6ee32fea36c1f956e1a240b34f56b6ef06dba6f9ae8883c7df0084d2fdc5aaf1ae0b127c599072a43f709f769116404afeb556c6b30c220d0c7ce6b5bca7bb49c4208cb69dcbb31b1cac2def3d09550d715dbd77b7e0fc33eab6e8bd76cc249242b8c907d6bde8faa988749874ec916b4c27fcf5eb7a87dcf08200da2ccd00b73d8be50f4104e959d065c45989bc9bd20e277f26cbea11ec8b2e79ea41f2e1ae447a518ebe9e511b2782fe79821632f28290707908fc589a0a13c1dd3c246d096d3d09430ae8b4b672534e6882c945c24fad7a1d03909b3206f82c76a83ff43578e60d4b116c7749ef4864d9987ad79d5bd3c9505b80e5d9f828ad9b47a3c7820709c3de67e7ce236b7c8815aa8e384e50195948589d43fc8d41cc457036d13c737bc71ec5b04f4d56fd7c842c86ed409b4cd95462891bc5a85a7165474c42f857d850ffd206373b881a39ff3ea005a33315bae2546abbb413cdadaeefc61f19b1b888f796bd6d8a01180c4e8e6f5ddaf3870525c543c2e472ef414ab3b2eee8784a39da691cb3bd9c274fa1ea99c7e6beeeae52f91406d6307f3cc62fe8e76fa9417305387e63dddd2a8977b3ace6510114a89b15be93ddac624be71fbac63f5ef847b75dc0af3fa1aef646c5c0d6490546135bcf3243dd80a71355ab2c144e8d8745952683a2abf4568d7b0c4865cc44989114561fed3e26e270c3d86460509b362f6dc2c312c07eae9ac8355aef1d4465c329f17bb1ea07e8950eb34c93447ef9be6eaca4cf29098405e39a2de3a92cf630d3447656e33316c268504078f66d9b9998a9f819766267e3714053103c8a0042ba9d9775cb5783a6b29b72b8f4cc0d921f6d59a7cd187eb4efae405fa0dbfcf4b839685d2e386b9c23e1e1692eb911de9803b212e7efababa756af592268ffdd0f1e9ed5c132c28defb81724f4786ebdf6b5b3c063e3a1093a2d84ca843bc17a2d49b63450d63d408e647b627e992d61ae8c63ca54f535ed04ea52184e2c7b37664ce7d3381e4f14df9b26831442f5da79bebc2adbeb9cbf60bdfa9e0e00791c2e0d3a9d939eda1ea4445cc691de73925386705244a0f8753771f4377c075179c9f26b57c1b75b1fa74b1f2e46c885a8a61e84c5d86f608a0e56eb7302a966d6dc999618dfc4dce398195b49f835cf8580cf5b846066eb1097fd1e1df45776360d7f38081ddcd6a530bebe287edc4bef9b298e26e7325182f2624429685df00f7fda2a818ad7fa5e5d7fcd38e0b8b1225c1cb1495c29c19bff585aea676b486c65ab362bfd517c956f7fdf6db5c41042e2e95b71a6a42c4ccd52aa5282fbffac21e9d1f2ffd550f942c9d9b7b80f6371f4ae04e443ba6d9de8c2ebe527f45e7eeedb68f886ec6ec4b6b9b48bde97af338e2504e7015708d2b6d07e49ad0842f80d341a9afb12562c1bc6088cca8125f159002c6021c435f334ed5714f3319473114aea722a78a933c3135cd3f67330a804056b90c79d8e3d5de2356129ff5f55ababc88a2ee7d0ce18ce536a8dc4e8e67fce37be447edaaf2823afb629bf12117d8bd1c3e0c76b83f72163132e32889e1be0d02f0365d11358fd4d0cab8329e5cec0144bfb94ddb462d9b13516c15276df73656224556561e9f2d3b65c3d37a9e0503c04c4866c9e61fd5f3e08b3039854ad4f55aeed450188304f33aed6a7cda4d14405ae8b1801554a74617409fb5c01b97e456810c5fc0b6171003d8ec1ff797847923532480f31f1dabb62b265f2852d3412d4ca54325cd25e10e9b0bf6e7f879be900ebbf217de53302ef2207719dc21dba1eae85fa1e1b91f2d02ebc9c266b4fd7ab337bfdf497f8c62e04c263cbf165951377ea932ac47e9bcb836b1777d485a5d2c20cae477c3b4e257ee90fa5a193989ff15f91b44f34358655d6dcdc307b27d5a6b5c9fdc70d0ecfd039816e052b960ff7719f050feaccf1ad6c815f4c77169dcaa1b9de76b06a9ab45c103d808396f9bf843bf2b9f01c5f26f5355697c1d6d34d2cb1e3242f554fcb4a4bdc080feb80a71f519cc5d1d0101f8db78ceec604b1bfe4d38ee68787d49bab09ca95e7730e484e873b93d9b15a695ecaeb3ffccec381f8c1da0f6485a6af95f590e4ca6fb3ec2f7bccde1c50f770da9b10a81a89bec889d71dd6ff3e33d320b90e6312e27d682f15f885ee8855965af83099a423a12266b92b3fc1c313e53344c628cfce154aa3c2028b5f29a2b2a6bff238dfc420d8a135cead6457e3294bcbdd33d20c9520313a63c6a42d9475c0858027c86cd2f487d8b36bf40cfb57bb3adf08718be44d5c07d58f6b4b51d39a0faba2a3bf19a18f40c3200b3b8c76c1e771d24c8ed40c86eb17d35e9c4d920d6aaa58e564445295148ec6baacf18a074f7619fef93ea287d9e0fdfe93510285af0ae6e6c4210ffa817a779387712661ceacb3b5624d2d0323d5054093e4df9f92ce664a39ff37bfbe8eed6913e22c4308e75589490babc938b51ba36ce773348583ece9675ad931c2bf191c4840bc3a4c754b848990bf70e47d2d2c667924d4886d63be7739928ad9eb5e030586007a8b2e22c317fc8fca720261f410461acf3725cf496702e1434db0e6fef297226339eb6a0803993f7554b6660f0fbeab84c26282a630fb9f5834e148bce823869293efe7564cd6152c24339dd9d242dd3cffd7cf81086695cde6d3308621173e7e6d267918353d6d2a478e6b8ab557ec5c95b883997575802372baafb13ee92788d043519eee7e2a66fe81d324de0448dc45e222a7b8b05d15716fafe867335d981b4202c1154e781ed659838b9b2911fabc19916d4f86eb1083fc4202148d0dd707cf523ada9e66c0a16f7a5de5b95a783e308255c8b258bbb26d6d4e7c4462fc6f93b82501b0db98e81271347e9e9f2626eb2737bd682047a1031f34d73c825659f8719d26bcd1851910db353b5f49ae17333d34d3ba7eb2ea8684fbf5c3a7eb0f487edf5cd410553cc77c3f7302b0da546771bbb3358ab9681772e551d1807119f41468466a92425ede0b76fade7272793021bcd7ba8d5ea91f1ffaf1e0d5c3a810cf1103e61c867dd7b1283432b67c4d6e0c6a00187c868aa934b4bc50459a5abd90cf48cca58216a64f2b398e57bed5227c3086f82121a6506bfbf09a689f7607348052c8fd77d3ea55130edbcb7398003b7d39ccb9c84f46cd409515432fe4a051727726fda7c0143f37f103e078c87c3d781ceb8ded32c5ee58d0a561ffab255fdc5042cc9baf8b254bd5cdae9543e2b8331941a00275187063159344807eab519ff6506b902f2409e164d0608bcc4a78dd32a81727742b4b938dadb464d94c0a86e3fe6fa4fcb10e66288e3244e7b6cd14a99b3f41be87f3cd70b05f2e39eddb7e3f58ae9c8fd75f1c3f0488c4ba158fbea8f2696ff4e7d821e73995b5ac95adec056f0eac1db729557b4aee02b3e151d355b085c615ee38246041a9b328c2cee2ef4f9912170288363e5aea9709bf14db2272292be14556e5a43ec6abbe3f4549ddd078e9e0beebc8a33fb15fb5c155b7eacc493a05d30ba8f6d57dee056fad4e9577aee0bc7e95eef7f345f9b1d39395d94681f746c6cace44d3db10bc554825b27a02826274a699215e326312ffe569b4a91b6f83f315b97e44d3103b987a424436479edb03ce2377558cc68760e52fb672fe4d5f60efad1af301c01ede8841c35d59a08e6b8c850b1e9ec8de2f1c450ff5e8f5a09a61c1bd6c807dc1ab89ccfea68c637c22a547bda77e17a6b132021cb3afee8fbed6dbcdaec5c67399a11a08f0c19f36cf9504e9269ea014efb2b7c5d9c8229170d3588876c043ff7a5699eeb2d8ba20f80ef350317514d12e6f4082cccfc3b0f2ba35060494395e06d8e178e06274e68c2eea6f8be1cc7dc101ada10b853b6259356081fe36d23c0b184ef0b2afdf6d8731220fe2b96583576efefe33951ea1bf9b09afc5ba92585dd6f82a7c6e0d2feac550e031ba4ac0ed301df6aed05a590c7d415b7bed7eb1c3e6066c850e3775d2c1e1247c73e75838bc3147a397d75fc4ba27992f7fe2a3965c8c0231e3ac39f69f903fc6d547d592342806b9c9ec93a7175c6a96fe5d3cc47d597a59dd73034ad1ed92845bce7b60154ee6115f81e6bba6bd1d05f231c2d149d158ed18edd8d848c0dd611b0a502ff02c8d0a081bbc9ad27e9bd7427d377f5a788c4762d3f618ff49262c67579bc09bcc9deba13992425f32154c3f869eb5b0ecfe046e4e3c55ac8316c0696b093e9919e2bdea7bea6ce7199f043085a8f5faff01addf0dfd90871ef7aeac6ce4a6caa29d28d13d8567dc5171bd92398f850c192731274cf04c33da2415ca47183ac5b3296a806e3e5e60c9ccf2b81e73e358f4ec97617d03fa7e54637a61eaaabe4c1a37baff1b70312a99c2124fdc91d0592e3790f51bd17db3f37a23cb192381bf424a681765a703d448a9fd342dedff56721f3fce8f607f91e68a10cb4d8985a72b74f300b43a3c18e7c2c014957141f33de182d7e4107592c9178cc33838ebc21c400570c98ef14c896e45a7fb60f2cca1750925b830a6083fd22e373fb665e427094ad15c5fd4ad6476ebf3b15405daff8edf3531a98473c7a6a95ade3a77409bb6f74a5c51b29c55c3bc80120ec1b31a5b8e374a4d593c93779773a9667f30063156e87ca698371689cac3c1c81db26bdcbaee64a6e69f773d080f6e50fb32b6359e82ce78cdbf0a92edd7a7ad162551429bb1224f38305d133841a6df3e6bdf3b3127b89a9f578a9530e291a0d66de10ee9f130205fe76cd3e820c3e89b9babc3dccc857c8b01e39fcd7326e6ce79eb3cf18637931dafbe41d6aa9dfde90e8b9263214a170edfe72a6f7836a1b207b11e84156d71306e0569dd837bf5860b0e072f65cc17735306d504c40d4484e0b9bb41a5716588b74af05be70f1ead795c86b067f65464c0d9dd4d0bc4ff34e1ae5cccca5ba9cdb74835db3c2abf3eb8d1c4db220358192d3e786afff8beabb032c0c136c4308bc6e769606addf03e436b7e3fcee7acc167c938087098b4980349ac9ccd49299e28b277a03d8b28bdc23805712709b120703b32c80fb1d768144c1537f93872ff0155d7066b848eaf6f16bd308459c201b1b2e014e5ec1d685ae1870fce664d2edabedd4c43686d43a6307cfd044d23dfe29e05ee4009596366153b3de61ee9640d66798193ff89c0578eca3f5271e940028d1590c84976f74db538b9d3caab786ca01c5e79e7d79f153f2631c1925613681743b91fc1c3a825d727068f2bc7507df179b8225941d6b60994911bf54d597c1fc7d85f9396ce13fac9a4c0f5385c984f869d5c4a17e47fbfd8514f7dc356a57e66eda4322939ffba267a6dfafde9d5ebad1d632d7c768c9248bc0177ac6e8f60c3b9670739e773a638e07f7ca81b7290ef5c1855f9fc7a9614beddfa180d303c46969130896f2d72a1ee6be1c16aac5ac195678b76f962015ab245f4401ad2d855aa30b7a0db8d2dc378876c2f55a76dc83a3b92e0fc79f089ad5e662dd213ed3a46e478938ee962470a7add877d107ca52c27f2c31c25ad5dbb8f38683ef29fab1f5a62b714e616ba93c978652c98cb42089c2d872c4cb429d4f2e17d3e13bb747fa41cad3523244b991cb2b18de2cfb59c58c1f9c13db3581d34e5d5d48950ea5a9f78dbe8236fc095800f70245bb5fa2cceec077f54e7984c8c5b2a236a042e4222ad5884b88f9c910b4daf97bb5cc33969721e53cb6845cfb74382362c0a837515a4ba4274ef2a4d19df90bf25721d7607ea94712b035ca4102fdf6b482c7293517d1bd773514adeb46a8548c6f8bc324457b25e066649a569b5e727af00c2414251b72f8502725985396315f58dc10271274002ec7d91ab50f7869df3adb97e1bf554e7102e66e9ea560eecda8ef6205d877547318123537433cb63a10fb5d268036f6b0a28827665ed3b170be428acfea6930858e91aa5526934c7983cd99dceec8cd7b6225c7959ce143a60466f57e48878dee0522f644e07ced054326ddd721817123b8f93d551385ec65082f231c608221900f7b4b06203f46b195578966c8a9cf9cc7ae6fd29a7722b9f12066ad33e5a5417bd4385a8d197a25c91b6e3009686d87d7db1c479aa37e0f1098c007b79c8bb4249f636a2f885108f2abcf51f67ee2cc7b11cdb1894d053520ed0a941bc6b1291490c041c05217444d59ea0ce15f45e3d0ed5577e77cd1d25cea0a85e281f305886bc7be6018a983679ebba67b51fd926d0bdfcdf462dab5d8e6af0e44eac194bbbbe8ee53bfaad2bc0937d25e1a7ca37d5c353a964d37b5e3eb369690ef9eed630c46ced958e8807af485209012506de2ae367b5dcc89147d83b35739720755fa4eb07e44b66da89586178a8bad1ad7184660b926d3ad58efd97409820cd8708f85dbe70a92df63d9961effaea661f394c61308b72eef1762b473bc412b4e983abaca79c7693e080e7f69ff9a2361a27bfa751cd1f1f8d1b39648b2322a2a9b82630e613fda0ea04c901c0daee1dfb319b0c5760ba57d03248897afc038b0120a86bed11889518eb3f3810ff54560b071319a999fa06d5e1201e0085626c1dcec0aa72e26c0f2523c77fd0a860a8c0aa5b127cfacda62c73f38b4f9e58589ff49449b91cc593810a5724f72f4cd007d01005808e93cdc56449884908ce0edd6d2a2aa3f877168b4146e793dd376af524e1e46df38b873c477604ffc111f0dda319dbcc078fce2772228d5fbbb7312c9e269063ae8d94b9e31ba046c1eb2f369b750698055377e47e22788f751b32c3e1d74f8e8ea1655ed75ab096733fe304891a186aa1d884b6f2b31dd24e5ce0b0992613cfc8fbcee7ed4c1261af09fb3681670cec3e7142298804ff51c922f817299b5f3700e81b6557d56798c001f5fa45aab03f6bd4f8e5e83687e43c5c2e0a6690ae348b5be55c7b35dc22901c2cbbca9cc79cedd9ac6b95ebc45b15b5b3d7098ff8589686e1cfb14d0a033cba6902b995cf6e1a69ec2a9625440eab0f2031f3d6f424e90593a3162805ac728e4ebcced1b498c27068c3de76791f9663eac54472bf3e8f27a1cc7255bc7995f8d8a538f455c36616d536fafc5d371825f21a6521e3352c8e3888a564e7ef0a086df26be9e92b67ff0c786e04cae762cf3e62b35473aad7d874d06deb51bb252efa36e0b07889246faa6c073a32ab0fabd315bb7b057e95c64e6cd25b070f9399608bcd2aa2dffe0780ed4546ab468ef00bd2992fd4baa29e6c52e440e8ac033a883a2d1e63d0190dff7bd1b9b93f2f3eaa7476a2c215e1a86c02d9193339411cc47f574e27c69452a1ec0a4420f21ed62d5c2308150ae6eab3c2072f71f14bd8b2aab68402e02b8a7b1c8189ed2592d198337a4e2305cd1af3c694bd2087bea7b711728f41d6c59118076878c72d5bcf0321c870f36afaf6bb59231c466af3039b1bf70b9fcec8ace29438df0ed026c4af1e1b15989c6e896217fa70469a339a559c8345b1e0e4e44b6a1b554866e4b9cc660ed1be9ee4d588c23b1489919829984773e596a68fc7f9efe7aa5b345b3bc08d0331a3197e92304f0615c107b42fecafc4a3a5df84f6e47a8e9bc91d16b131d57810cdc0f2c665e81365d98f26031d33a48883e932e0a4164f06a5f4004c8b58bb6a1fd10a2fafe2a9f3837ae6ec38a086467ee74ba3c5b554de7f7deb07dc0a315a8e724f13fa16ee8a0a9b74e968534d8b5f4d0676a517eff9b9f89c25390f0a37a813927203bf506bc274573625047e93f03ab2a6f7ddccd9434f30520bf8abe4bf2be0d70d2ff55bd2ffcd9333985011d7b15c40d70b7813557322a649224bc4dc61f5aac27bbd123034e496efa544f4d22e811c5564f06fd2e5d9d4a00ad1ffe598410a0406fb571339a76f247fe80ac0fb90352e7da263fbcd7fdacadd775514754fd3f2a8ec126297efbc3ba9ad828344fcfd4cdcf4c50144fb2b27e2d4c90dd411034b382202e774d81f3701ce8d1693abdf9c5873bbc0bd7e0e086ae685a9aa90e290a6640d86c06fd98f3a808be9d2fd1c2f57cefdd95fa51018d26f8c0e41f89106dcf258786131937e9443e840be54d1fcbd8aaa816d1c9aefe206e7fdb534974b6e33eaaa6d98772c288c60e498359b2e497f12daa13a82c845c6f47edb884bbee04a85eda3384c53d6f93f004c7893e56adbc05f8ebb107735ebb98aeecfa4baf8ba8a717d44dc96b47c9d8a22b4b5551982d56667db2ab24cabcd009d46aeabc2b31f2944a18fcdd0974f1540bbadf5ef776c31e5342202d4e6414863e1e0d4a5950eb6df7b6e313827720fdcae4083e4f21dbe00329930419ce813a2d6eb950fa21df8407d297810f063ac21136ce4d50ec30eb5e504216f2038ff3e896a0354d6e8157a7697949c14b9d7dca22bd8901323b5afca9c5cc98e78591fe541fe17008d0492c4399e1ef2e2988bec2b5d42837b77157b7aa9f181739470971fd528799e2726537d406a103a7f999aa13f4d74484a3400859d1daab513daa6f25f216165a7ffb1a9931e7706d5063153c295f7f029a85538a36308387106d69b6615f0aa3c16cfbc97540bb794e7f18283aa741ad17063d427dca3869169d936983bdeabae4946cfb4b1c5940d35fb89d7b0302bf6f24c219b5e191c0abbde51703d15121ed8ccf42fd784bea778affc243c8fa5451a732da0e5b294486486bf6766afb004ebcedae2df8555ec057b1fc3e4d5c951d71a1db10f2a668da2a61269106461543217ffb20eb274bde33d17b1337a92c13f8c9156e08d5d6d70883c6055c1a5541b803782b4641bf673fe479df011e5ff83cf0aa79c765b0b899ceea63e91dda1be65d59f292edc0b89c37381bd9c66db8d72145426f5720405b64f94d2273d2eceabdff5bddb78ce052edb9e80690937d3af729197f84dd1c3fc10c029b44e9c32965a883bcc8e26f92fc1ef4434368460e43d443dc81f7900f7d60a93447a723fec727101ba3c50b6fdc9ee43fc67d88c462ae2302279bfc7567254717bf6f12d78ed7cf17a2cc0fdf62782127ba1d6b329f03b1e0492acf9e2dc46d4630ce243e5b7452695079a1cd3351e98a6a5f901e0d01bb9cb19b47ccc3b813c65e15696cc565980ae31ed202b559b53613e68c4679dc23a446845e4210f664968e042ad6c3b924bb104d4b3f7c9c6a9c5860d9af01c3054e6ac5cca3375ba3b6a2736cfe7d1162ea203e37b21f30ba7d4fa7a245fbba7f293d9fc9d4d8ac9941ab14ff3c63c09070b9532c4f950a6fde217ec08d6c20dec09733a431293b7f1b2071ab000f48020cef364e2bb0e380348ead31c1f98dde92f25ae37d34499fca0ccc42680b7c14b9ea158d9f7726d9b6a4d3a44f33aaa32ac94d5a41521d9f3840105d0eaa555325c90161b4e6979532f5e4147a6f9522377e1c78c25d6f14299fcde3031e55f33b3645306b38491a2ce206197a5286164e9214a282dbaf37e2c337e982099f740cc14568050a05ebb253953bc580f78d2e5f2c1bfac9e5daa4aeeac80cbe3b02cb6a4c4090635ae1a1ae45a05bfcf8aa7262f3577dc98726c7707fb4ce358cb23d645332c4f29429c1e028c4d81c21401326ad2d99ab9fd4debc2bba36d72c4a19b54e00ac813d1cd12472af59a2eb706766e7b51e11db58bab3c0ce6dc354d531c3c5a19bd7e43090fd38765b7d6d8fb4049f4e00a68af191cfd61864b4b87744953bc19edfe8b476983949684575ab0299c5fc63cbc0226715787b84a6f82d8bf1564bc86762500606b49dc8f658a61402e546e7c7abb63868feedaa16a901c9f99b43cb1792924810a4293d944e8d716f23b8eb503aa8c9db1e8478a81710a7460cd582b0c7be564eef73bbdeca987acb5f25647441d93f783d4d0394f429b1ff307a4dcd41ca21040463c90392bfec7d6d3f01722e2e2ecedaa0e54770fc1448cabc5d0984574fc67819b1d5131d5bc07503ef24b1a1608f7cce85d0758fdc633fc525f6efd5103089371fc4f4f66e5dc631cf8d540575216f3461826b681d7c1d2f59de4d0478c2dc5b9ca111174b0b9743bdbee46d8029a57a98c30eef465f3c147db200c1465e57c4d1a6cd3e81550761bbf03f40c272055dfd13f5096a388deef95cedb174f9cd892cdea6f117d6567a39849f85081088f2593da01285355c41f971d60f8074956046a8879cbac2aad2a698d6c3656b323137790b2843f1ce6449f775e733f6814f7212501dc0997ce80f10716ab21d94a15dd02acde2ae88b76d62889839e842d59994ebd76a0f413669641d281b78238e64bc6f1611c9232402f428ecebdc918a73ee9bd2c2c84730f866cddb635af4a1ceff61f5563c304d67b732f6d61b5319b08ca75e60a5b3220da2f170393b47c747ee07f8e30486e636572e791406175c7018084ff0a28531cbc1d8e776b9f07abd9693fb946f9f9631eb9b8a87f5e8be25ea5a74c3e534830f3103f07e7f1cf7b0304e1a5f3026f3d658ddd5183a175f4af80150dfc5ab0a8c381afd7cc514000d4a3a9a9e50e6721d702c5d101f66ccc14dd6309482df50d051ed2b6012c80f2f8fa98a0d1d29804f6bb2c03b2b5b5a8cd540d7ef1a689b937c5a57068280b4cd9457c5b7bbe51d3dd843f87defb10c4296f4c26f974ceebb2150732d65d954da65fbade684d4c34e5e07300dcac2b06c1d1f964582af72bf9727e7ceaa8c8d8dd62d0dfefb6a8f1d6af57fd7a3609a8a5c33ebf4907bbc39275cfdb51a5c17d5048520246f073c6ee6611d7be30111c90f5cadb612217d19d9622f44d0620369d44847e4d435ca4637f39d039de9012a47fcb6531af38ca35354cefdb0fada257365c49f7c576c99ae31afe0fae521cddf954983212f9c9661ec6d6954d6d178ee4c4c30f458697aed2f897a110babb03b4138aa04ed84e80aa674e31eba1e590a7068b150080f0d3ac253b77f97367b4c5ded7e34d2d997319c178682f59f5227f23ee309b033acda2aecc8dbb9b31abe498fbda98418b1114bc9b38f5c933bb9c6b90a6e0d5f934e70ffcaf153fac91ef7b3bf8edb895521e1528b8f3e04d3900ec9301f2556683161f0ba80eb468f475b7cf89fcb7d02df53852112f7d672c55c982710421c353609933d267f433f376f73435cc68c20abb3b31f391ccad27d4174e4a90f38d36801d23e9557592f8d502661ae8b737dce02a3fbb6bcbc05acc5d5f77f0760217292ae766bac81e0fc254378cfaef85c846d775e051c9b08778800007d66abff791af167c449e017671a2c116447dc2310dc5bdf65b7dfefaabb175926a09824ade401168960db218d7809e11a63872f1d30e3ef9108b73ce1a1d387a06600ef867c6de4e0593722c8bb730262c75b5b9f50dc8670e129092ddb94eb3a824f8acb97c3722e7c83344fc21f65aef0d9f4a0aa6cc16d732ed0ba666cb4ef4453606185089bb94270a1ef8f4417ed52ab8d8ad06d4a8d62734adb55b13fcf15500cd70b521615f94847e0ab2bbe63ec3500de453c668836179137b57f485e15592ebbb9d84e6cc7db2a030204b7cd65f6b512b5c430599a4abf03db1a8384734d5f9e91482cd205a04758d1d66e74ff63f61f20253e7bba28b4eee4e08242e09eb122eccc7f757e7dc6e19204adadd29295efc724878da70dd41e238861ee76acb2654d1a8c6fef299d50bb44a9572ffb2a11d00c5e55e3dbe620ab0f4d834940c440f664aee6edbba8dea88483c6c73bd5bc04ad9bff8c6ca7c3a34ff1e126226967e8c804f45eff950f52ad9e4f5a4ac36e359fbee7ac73c001ea49cbaf2eb3b9b19b2a936d7a8262c5d8829419bc64c831bf59ff72307286d1dca02d539e6d8cb025dd77c5cd18075eac3d450247820e87f2d256';
+
+const financialImprovementId =
+  'HorizontalCardListFeature:f6b30a38eb547f1f75676ce0557a7e05bed66de665564bd0c5e37f0692676795e12dca024e16de4c6d7782ed0288eb10570bc45a488bee66c93ee7b8735beb76978e3501e90a70caa695d63b60072fa838a9bd3f1beff48dd794839ed66b0fc1aaa8569f695dfd7f96a84d98eb72b535d107dc146eb20b7307aac969c622e54248bb19ca02245f1e6f2118d072912ab5c92db0787eef94e347ef2f669140c7f6bebadda961a8982f5a34c1408cebe2153e896f24a9fb6bc5a4b52d6046fda58f348ca5fa395d4f882d440e91a3338fd7f4a517750395be1d8e554bf871453b97f2cfb7e90f47414f1b048db5c5eb48ca7324f7c0da481cd7cc98d7ceb9add7ab7ed5a49b7d7fa3b1e70f42f7f7301ddce08ccb9e92d3d018a15f8b84073f3219350d7a2298d5952e748d3de17d487ffb76137ed05cba21c1642b32432fa3f5cd1334beea264ccc93ac28c9c6ff472997f37656fffbdb45c00d11797618d405ecf1c5872fdcabc6e84ee6a069564c0f8d01515524f4b50b159526c702b096e76ee364bc9022330561b0955a9bce83d5c7a649391e5c37bb96be44fe5b40f1e2a4243cf3f4c2887970fc68bea99e01536136909de73de4eacd60211076f6b0ffe043e565e1fc2383daa46cb52fad604992de1f3c48e21e16a08cfcf862610b6480aa4435af3a60d1d9fa584b13ba757297ab370d13193ff14874f06ba3702511000fd167867391eeca62f18ae9f518bbbce17cd16b024972f9cdb0a4627c03e200de59ae58be6a0e344ba7a0c1e7e5299063d0ef1b19b9559153e883b6ffa8a692e83f1d57fb6e5e11edf77af65bf9042f97a9f8385459a24a5bb0efab7734b22516f6c86821ce52362a3e9b87594c21bd1b17a1dc8041ee99f2fcf84ffea8d41d01d6a32d9425d65c0db34bb59df79b83a006b44a4dfe443aaa20ebd208a42557b85e78a03077bc0b07a9839d184c47cd302200faeb526adc12b960fd86ad17eb612e86073f5f6f84e7338c312984e001dffd2e2293996d5a79f58bd812c009b223f523d47c744145d84b44476e786a8924eb1c616b3e408a82bad88d359a8b28afcdc7c788a638f851009e5be2825d275c5b25032d368fde12fed8fd01414a85d31b36186230f8fab68c08ee845cb9459fa8f85575467a84ac8cc72de1aa623a66db5d2c0a83ddeaab740f0a2eccf91223c7a6e589600d8ff6e50500eed00d76c41edda447e209096f4db76e4bb69daf081e1d23942f137cb035632ff3c32fbc835c79fa0f675516639eacf232298ebfb932945d43675e7d8e15b8485132e1d6aae437afe74cdb93329122b5cfe49c7e583a50c0313a5eac59a22629d17081f0b30f68d69f03df6b9a83eae222d8d7948cf060f100c28988b3bb13b7fbd9b0869bd0536332293d04118f52f5ae82ca57fae5729945f597f9d5ec4dafa2dc6d6eb49fbf8b11cb9c7f0a798bfda302626bfcbfee973ea549aff7a79f617be14da6ddb9d236a231da7d13f79a5325c99d9513f55428ed31d5d2fbc68b957c29bbc864a0e2ce32a4f6e2229c36fdb73e084485e81f2db97b26c36b66833877b0bb8ea4ce3eec032006ef73480e77eabc083b028c3626212f20f2fb82f06d09cf94d6e039270c4da65aafd16e66e0feb3c53a59243d5daf2e8cf5e357f4f1ec250d80b2555923a5b70bccf213c95f14a7ed65843c481e79c39832f6f17e1526ea8711ee829a9bf39327373a3ab1133395f652a590fd2c477828deceef544eac1cd607801b1dda01af7c721aa582438caa8bbbedcae2b91705669c0d7880582b4f6524e7207c11bef34b9b2fd429f796621d05266762b5caf9358e9a87a038bda1b9180f206457586428629a93ec7eff65cb65fc3828c6d53d473dc3f89aad7322acb392eaf3abaca160279c41c081ddaaf821e0a886135f04fe962629b1b8c4db9066517246ce909933d7e1cf3a0be808ba62ef8bf87512d5e24f79454fdf0f7e28e634a1759531cdb716a8450f66f29b38d4035ec87bd27ae266427cec710c0edd547c354fd7a9158345d76d1cc82fe845e662c7f41284bbeaeec358a48f9c066e64a87f12e4195a11bb83d4e07fba4d537b67e4db60721e0b595ebe97ae1cb4c767be62517c9ad0002b23f8143d8fe26af08a3c2a1b8f1ac5f11b3bde119b512bd140c8f5f089f03bcbf66c40eb111744ff7e77b2cc1136e2f4c88c796a47690b6774d2f841d820e40462e07837bd1465734adadd0c79c4454aca9dbf3d8aff851a46e7bdf977a5ff14023a83198888c08e8f2c27d4fd8381ef6e8bb51e7cc32c0e19c2a0db9e0e14d976bc01b851c22c73fabd48ffcba37fcceccf1ae5e32f378e5a238ca7fa12f84508c633db54aa8f161bb395dd2fe657a1d0716ad79a1335c1a246889783e5bb7ebb6ca6e0d005f4738a8e0de148b1d0597032b4a3fa5a5c4f378b14a8b3b1dfc4f3b18dfb06843d01ee128de4a904ece14364a2d71e46bf46194eedaa0f13ee0593b468a93dee62d0aba6beff7355c83aeb238417d853c257afb78e1b070de2373b5fe69930a5e79fc419fb31ba2b3328b285f50579cf516bba277fc6b0af2b6c5dbfe02a2b62f3f47a8f6a330c59f857faae5d548fa925fcfa1720529e3a1f86c4efe2baef7c73097abe378ba32b86dcdd28e150963a6e45bf03c2eed204cddcebbbd11e960fe6646b2dabf1d89ce24cb07ea769bbc3955f05671c21ada667bc259f02f2e61b862903298baae6927dde4fd171cc99350fb4518466a6648781103757e758086114c8cc0bdcaece2593fe49512d6ba984a75f86c238d87daae765d5d4916f34763c251f5b0b3f48ed1c81843e18b6256935b2a685889edcaea2cd6475024379a9b2f89c3e2a1647befab98574596a25ae66f368ffc0c2a38f9b44d0cce6f4f92911864429d59700dce64e0b86ae93bfc51c46236fb46fbe459b1380d161a046c5e6295d7023d751fdeba30326960b1287baea7d15bf49537ee8bcd03dc111911341fb730db059b74437a5c4684abeb13d60e9111a74db06f489469763b8a5ce7a3aeb028f47dcf311d50d097b405cf2040975a36684eeadc8eed381327873746dde08ba22f847a08d4f304005169f231324e7c087686d1b720826192cc26a487f2203ca87001fb491eb324179a95a6b17c14a8cac0ac74e123ecccddda65b30d66bf59494c8997718d662bcf5e8252d233aa475d9983ead74efa1f1405b28aedb5a641ccfd2c0251963580d454e17521489e8fed184c895fef60566e9eb300781d775cca3abf5aae6c420bb2b6d945a2196601d8bb877d3dbd3babf5396641d54d62df6edbedbbf80d66adaa7e505ca34cfb176c5ca551f4030d610ee8cb98c65686c49bf062e654c4f44f8b1ba2ef14964827571b06b4a625424b480b2cf7eeb95eb69d19a34e9676bbd007b22b801f489d7dbb1e1742c0c2d2d142fb9bd89bc77afe59f4fc60b094fd47f4ce967b35d33d8895393e34f42d25cc9a986ff5aa6c4eccd874b0699714f4fddcb7c7034ab63e4c15fcf575f6943dc722921fada0846c5e8eb73eea5e34b559287449abf6956b5eb12829f5127c03746f6b12dbe00c9f0bfbbd5f424456c4ff528abacc14f79ef638b459568e8ecc269928d1bfb781ba0132d9cc13da1375eda89a166271006d5e61245459010d181c0fcd592e84ab8e3813bd24c1de31b6a5988f179137659a7b2a5cc37abb32b6834b580c2ea2de4e7136e8d19c03c7d02ad3c6de752fd3c8d34d36a5abe335aca06a9f6c7f9634c689f7e3ec88f49f9ef6a90c08264cf0c1fe3a876126b51cdceffdab5c4f9981f7a9594c0c7db1774bd4c918fc70e61e740cf4d965f4c911b8941272a0d98767ed876a9b9c937855acc428428e523acc42f1fc7dd2e6853671e6fdae523d6c9e435add7dfd4217b287c801d34c24bc4fe8af775b7438b71b2fb87d22a865d396e387421b1464d781bda566babd62f82842f760270886dbbebbe3cf9219b3d46af8b1ed38644ca8b7add4f0e99ff8a2f71ab1f3e25a77439608031a85bb6d419daedcad3e77f90678c571c2a87853d2fbf6ae3e985110e94ca7c37944586ab72a300b9bf53130fc70dd611fb1ed1ea1692d53130e8eb80fe051f65ccc631a2b8fb36b9913df2a1df9658fd000e080ae4aa248aa2c11a74faf7684c49c32dc2f8f68095e6e107bc74521710adb484d008df4d0a395d7c4b5b5d27204e8d14e8c1877849d17f3bee450f93034e7bb336fe9b3be43fd8bda5c38517e91607ba451d27329135738c5903ccb9e05836c079f54a59945d8683be4cbd416b6e043eed0a92fd94a46b61f6c68128b958b13367709e3a54d8b87e1fa3ec17abe3d17a6d7d26eb7a37405b095136ddf75c09c2e600345c433cf0dbe68a611638ad84d8914919f4de12786fcb020aedc10e5a8ec0d3db1ce59a95f7e54f5d88fb7f5dd0894d3988330683ca82d2f3d6c11d63fb0af4000d314b4f37bfc9a006364f941d2ca7f52c2b9ec81e54d2732ed494b1ed09c33d78d88676e02ad9c0533b849de00ae4f039de5d45dfac4e6228e179bc9c4324c84ecff2699c0669712d6d00c0d0a8094ea0fc482e5b6299175959b8871beea9c45a89e6de7da3c711f1a3b08142674906de3f75bb6b543d986dbf4efab795e1609f767b72af27fb65a6e1c0395cb5804c4f0943221b628e7509da8656d4c340f641a714e6e94f9d4244cbd344eaa8f9376a46a66877bed3e535eb5e223bd093dda68f7a839e7e845c0f01c19443e42aa9f6866e18f626e3a3747e19d41041cfda54eb4e32175da7cb0454bd119e78ad94e9f6c2171a93b3ceff09453f3e918947192e0ed505b3f6bc5c7cf7658228c501412a79d455c26738d7f0229586e7e2c4040fe8f1481bd2ccea7b4d2e4da67d51eac281deab03e6b266297b59bc10d59c7be8cedb429718dda5ec4d44aac48c574c131f3e29126e7b9713013b28e01919b7823860797e08eb6989b4f78fc39e636a3c6edc11ee7eee60cf2c3c9686a09c23011f2a8cc48efc6dc44f3596ee45e43897cc7b69f5b7dc1bad8009c22fcf5aa3cb9448d2dcdf9d00b2699396c268951e318e317a4d26cc7678101bb1bde90532f8f94d2d4aa9039c7f8368ab1aafa9da761e420a93a611969ddc5f65f167399955a6735fa3a274ba0c3f52241d8066e54c72e1946cd9c12e652a4de1ef59a052af05accb8e3853f47b028b5e6cb48f2fba63042e87c0e3b17b62e012a0b355972729bfbef1e74a9380afeb69f35689730abec7d2da8edb804c276c7d31ca2f17caa5aa0f5332fa138221c01661418ff240246c463e038df1f1e43dc79e6d7e45e7963ba769ab680986f5582b9379eac58e3b926aa00d7050a93b376101115700ebde2c5647df9280a63a926468f370ac56515d73681e028308ee7fe82e174fe4e40478d2779432ea51f73542d97338e04311147e433dd302ce1e70c345b1b0ecbe0ca9ad947d42d3f87a89b7c6a52d80aee257806326644adf0d0699cb587c5be9171fb798c666c542f8a440e2f0eb3797559b191965d5bc4a088b0f6b0f45f1d9240395c4a50f41e7cc748cf4a456f1d857a0a0a94c791d6bc2dfc6a2a8dc6add35f7f3807162ea00d506c625e94338b60e61d901f0864821c8fe4c1dd675cb05c304463abae530e37a2ef2c7a1e8a6d5dfa947142a341c98896a8cb9a353fa8d829a207b44454f709df632ebf035444ee39f20a7160c9fb5d0857bcbbe7679ea2808aed4593b86872fdbba210f100ad763c86987489370fd1fa27a00200ec2276f250c1d49bb2c086bdf8de8d8b2255319ae03215b851b5a6ddf20a4bc7524fb8578be3383265e999636a7342de592b58c6c71a2012330fe098a2aa9563e2eb912e187d37084b5d36e90def3bc55b00a977559be51bbf2cf17efca898406f30fa2f57d6b8dba649fdeeed7b2f99cee8333d6b90d30a9801f36f2825743e43bc770d42d176306653d9443b613691dfba15e334854b301834b4cee6b4fab81b849589f7222bc5d71ff48619d44cecfb815872feb04e18264332eaad758a2c0006f85e0a0821b682693d656ef4e88e2202981db3a680853321a0c2f124c295e362de26e5124c09a69eccba05ba6988666a18eefbdfd5f728c6d88683c28a437c26b3e92baca154ff9b2bc9a87e8814cc4f1627e07d6db10a27f3323e6e543f5ab93894b2c0e534debd73b9acec5a38f176a0cb84f730477de9b93a7665e37327abad15b0818e9b1e7ec62b35bcc509d2653713ac529ea0122c3656e6a0adf0ebc2a0ad089e9a7869cf7800568a7241afd4ec5e77ff37b8518044d02e236107b554cce2b8357cb0e7d79d20ecc12c1a72a78665afcc24f05b73a36c2e1ccf0caaca0b38ee733784eb892b624104c05425f38d994e37f79f431c869d597170c0b136a1abfe0f3d791002912fd54b9e16c53f1b1d4bd9f4704a12b19f746c0fc65c8352b4e41ef89d8523d5a40416b4f24e4b337737aef708523690114080086235d12e7d3ad34593e896547502936bc7a0c9ad90337701b24cf3d2e599e9ce47255ca285530ad0eb29621b39ef28ebde635d0c4f4500dbc923be26cf2676daf223a693bcfbeb4fbfcb72b1a879719fda7fc9eee39ac3737cbe5f503e69e597efe820e7b659a778cefdbb17bc248e66e114c0b6de19619c2a97ee9fd8faf4950a4a36e0e31e64a44e7ac466216ce8786512309439db136874a65e83ff43b31df3d877701b489dcb91fc63935f403c566eaf5273f09f7ef428243bdf131f5d5226ae462af378777cf70d55ffd58be9bb9c35097ba39d3ce63b3753aa5251959f1e05d82b407c153248bfda3079675302febc6772c8c1c42d50fb747dc8360b809daf83677ad840b09d1b74a488852f1d937aa583f34ab5629066146f5db595ccf7da50e470cee188abfb0828011a4c81efe871ad01cdaf1c07a3a86146b01ad3120fdc1b36f70a65ac6f394a70455e45750daf72bb5bedf151e7085611fdb921df2be1eea0e623c078c39c3fe09de205588d56e847d41d0b3c0ad15d39de16972df963875345856d6ebf24f29aed4f43fd66abe9c9672e17cf03485f5716b14294a015aae05fb7b0a6ad1e29144ec7c7aa5ac69b35a149f646100657033b976694cac9df3c4d6bbf4f0b4782a083d1cf8091a82f7092b39a394074992ad74fca33816d574f266b6463ff618db626765a8c760e2d9338252436aa50ddf7ee8f71298c1e9555905e6d26cb987d17a0acfebd08b816adc560a1918c59536d431e8f7d826c21ece77389bd827843fe866da2f2861a4e06649fabf16162bf667f128f6a193d05c6d93802dbebaf820ad5c615472b00e7a2bb3551ecdeee6f6dd11cf086abd2c6ca41b3b9642c6fbc36d638bbe283e3eed18438bd7d09bc303812950c157e8c79ce41c880e86b2dd13959499cdab2a37edfec834317d678cc96a98fd3e514889b290c4920238a7d3c51c21b9ff281521af98abd72d6fc9580f6ff91d20d228889f31929e4e53eabe60dec9d055f199380a8a4583afabea2be0765022baaea749d1c38a27422a693b0e548a6f5694139662b59d77cbee33782bf8268327f43a703b3bbcc89c99e50a2b61be1b81d3372dd8bf2388b484a55f663addee4bd4df3be6d1e554d6ed0c0cf9e22a89bff7a41502af8213522cbc031a5618968856681e9e90d8f959e5ca0fe8791cf73f852057abac5dbd7895f26af1b5e11cdaea4f8b25598bb23fe38ba378de3fa2abee97195c473c4c7c4678d30f77e69c8ae6b3c85ec7cb3eb04bc51a800249bde36d19c86e1f7068bda7874e4b3e283cdbd702e5f5db727f2a3bd7ea3fed2d18200c11db867168a27ead80eefec92fa083a96b8465312481076dc886e57a95c01ec653b40657ae26aa969391fd5dc0d56cfb51113ac568d00ab4d05a0908d695644ff8bb916cb77458bb067e97f804c560a3fdd710d47a26abe77c07b6d301481374b475befbb9a60fcb987680e8941a5dadc531bab8bd1f2f74773bfcfc6d179fe9361b13a5b988d7742db32cf8781c227d3c9a1dbe2cc66fd78d1f970f74917aac4e49661b541ac8660afe965947e2e99633b75730ab27e554eaacb953d791c82057392ee940182f2625055c766bf857af1afe715c39e18641bed7c7244139723a9932237a04ff4af750de9861b1ca3d63fd517bc02fc57e24bfcbcba5e40c397da6c7a4ae8670b647bda4441ffc4f9769f13a80dfd682ab098432eb7c804bce89d78a7a60b8d9078580ca24840874a62d11e8eb15f894c4db5edc134e7bbf89b64374d07a3b609e284dac9397e08956cdb9c4ff40f3c7a882e94e71d144f83c5b39d8bd77b4f249e14159bb84669ee8b3aa6421c6998cc21d450cdc2258c0da7acfa9489c0a999c71180ac94b11de82ad7e85c44c505650e9f2bac5536565bcc055c4a86a50b1e17cad4e9e08a22da8ca481566bddd281614aaec4926769dee8dd4ffe8abc72406b48a559ed49b482d367df46faf14096d3420d3d0bcdaca56acb49650e746e52b256f055382ffb91e9eb4df0b3f4f823887096613e13d54f0a75069fc394429dde96dbae6b9dd6874e7eb2b2d3d2212b5eaae734cbea5f3043e7b547e4ff0c33fcdd0f69946e802f40d2697775f6ec6fa4ba6e9a955bb932a2583d41e0cfe14524665432c09b899340cdff4c899aa86e40567b439e70bd1a937467656f40038742c58249e91a0f893bee9c2a4e8b3ac43c67aac5b1b31c2de9c59c59dd92a5fc6311e9d76b7ecf8e2375370f43427725cec4cfedbe044d52fd1e289f472c27cc35e344f645a607a9f7539122e04b26623858595181fe43bed52e19cb0c520407c02c325ea5df8e8c9e6fd8105de2cbfb19d5fdcb44f2d89f3938249a978d49815f667828111ae74fb8d5afa835d56610be3e68c4b9fff7af590627cd84f6e8ae822f507511f8560e7e034cb97e956656383929370f8d362d52f4cca6990cfb972038d17ffcaa67e805121ea3f492011a26d111edd220ab1751a8b7963a642df7406c3ee9965634d657dd0702fa891891773e6624e3c2aea2e269661945f8be35de98e17b21a0e52a2e2df6233274ee0b7f6f99e6cf9cf6490e8532c18209da71f97829269b4a9da0642df8b102cfadfa58f2610dd7c0b7fa76666b30dc62eb0e72ec45fdf020dddad0ddf6d3f7b382c98ea890a8250145ddb996dc4efd88067f57af56eb1698fbabc149bba5d2db75e590731598e844396e2c094ed15636f71aeb1302c63a71e935845e0bae1bf996a0a64e7bcded7a9f5f84647b71e30197918b7c78a3920d85a1d5160ac9b870dacfe1cb9f24f7a63bde4528b388d37f356ecc11672959fc630bc85ae2ba449bae1474ce035d0e47e43c0bde2e616d810f7330e95c58a1090c42e09bbc108e9da34e38455305fa336858aee8fe0c8aeeffcfa72ae8abc669974c7b7852b06a236869ac35c674744edad7ee3c3111ce844aec4082c6a0324d0b497a02445f8792099954b1cef5b5348b928fccc236f47c06dd497f4d4ca842a2d7934854565dbd930288998882f81a335d9887fe1dc73166685501434ddf8c19c96a8978b851cfc44ea06b6d967da6c2b8f5ae7cab1f8c07c31b337ce3148309f28431076f0c8385dc754e378b0a95027835421649aef94e935ae6b29dda797f87086cc6dddd8f31f69517189fc241dc901df7b9632de3139379183fc1ac4224c533673f4813a20d7a737d6a3094b5608aba86e4f8e8b6dd909e922c495a3df7cc631d1253e62b47ef751a7ced6f7524000fa16f7c812e5caa653071c836b7b6ff76682416ca86c3a733c20faab4c9847e62b5f6459f75a1f25f5e64d239fde8deaf873c8f907690ba794c6e17a48848b461454d6da5de0960aefc62c698162d0a952fd0d39b6b670a27d4c470d37915e4050823d2ef5508354c2b9a8d9c01a784f6d6d52aab09fa42b7a059da68a4664a1d7b61029015e5b6f84db9cfff814eec43c0fbc0bf31a548684122f1a4c07631786262f80caa9cf782f72932b75643894e3bb54e1f1566ba96ebc9a540ff039112640cdabb5a202116b37e37d170912840142a6343e645ce694abd82e2875574003e77dffb843fdd31401176fe35c4071629e1b977d6d97acb500f1655cd68fd771ff2a17a244a9616f72373158a1e53efe6bafb5b164b9e7ee0d5f1ea6688e4c517f365015bde5eab1a272f1471a102f95aa8efed5003aafe253fb8e228a3a4c10bd0dadccc3217e172eabd93ed74c2a6e5a799d8a4d78dd3dc9f80a7eee146925363146fa0c79559182e65943c2528d325e9f0189b1624e2e23dbe07e1221dc5cf3f1b478991ab6b5e4d410456b7be5addcfd28fa99ab544db88ddc448e282b3ec9bd289fd5fe11809825c4584be31a6182e0cf75a080984cd32d0d4bbad80dbcae3e06b30f5c4d145f84371418fde1501555849a203b0fb87b4ef35db7b3625823602d991413935a9f0fe3b2fcf97893da385fe2b6e90051d050ea423ba1533401826c81006c3c2cbaf48f5b500ee7c0138cc4fc533a2affe025ff0b8b0247815a5783fc66fb3d4396df705ba64f58de466192be36167b5be437d5abf48dbc91147eebe37f84a744872095509470ef62a36534841eebc539c0c12be19a57b6813d95bdbb62bf44039f830a1b7f036bc069f8771f0aacc781f3e4cfe54021e2e174c8530e3a8eb2988e2a3c296cdec0aa995bbd59d63bac8ecef0ffdbfb41a6ddaae5b1993f835944c51bcaf315744ccfdee2613bb3bd610401d9e9a77e63d81e5eb83393a9e22f904d6a5f37427541775d0dbf3febfc41033d16a958f949f82ca5210a70af765293e56063c55ce87ad98e306a2c3c08a172365d1e79cdfeb514e933a715e3c3c142c8458e25ce09af87aa7edc7e24644c128a88b71f7036009f0a2f79cf1ac3350cd492b792d702b3eb8ab65868d7065faa4dcb2b530d73d6334095cee85aca4b34a7841cbf1c0d6fe2be61d50ebc96bf36e9b011bfe5784abf540258c66614e3a16db46d22a5852beaf306eb8e508b158fc7e85e804dd9b660902b3d08af518d9d16a82460ea9882cdc42557ef5b7eedbdb37164406a7e56c94190a022112243c80cb52db81a4ba17b063044647d7fba600327c5369896b3ee7a2b0e774589dfd7c4cb752cb67743a7018439642140fce29b90c39f0824174012f4a70fb01f3432c54a33b8a71cebf2d1bd6befbe12c927f07459480454548e554cab2be2642f187fdaaee5a3332a08f897c0387d1b3ad3ff91a9559403b98193e22349ee092218349a8ec59b92a5879d8ea28ac57fd8525720f52e24e802e58e6bdcc3dd7024543e817cee1914d6e043de15e701252deafb517585fe09cf2bb16f4e4373043dbf387f60d7abd81768ad28477e112b20736c2f3585b1e26be46d2d3b70d7f9d83d5186cb91589dccc67933f5037100c724a14bdab1721debf9616da16ca0cc85df0b2307db058e97169bc2e34f9a76e4766205f7be3ae946bb5b7cfbe80021c394a4dcb5e4ee30deb267bff2a4683fd835f3b87851e623a6480dd7e0b59730879f18409177d4294485aad8cffae699d9a0a80ba26637458e145d9f204fd55baad615f32ce10673fd362cf3ddd2ae618dcf1998010b3fa53aad678966af264cf07b11aef5e7cca0fca6a79251f258c395a12f7b0778db0bd26153bf4ccffd4dd1f5f531d70050e2341ef85fed8cc17defcc9a2d78e9faa0bff625dba706530594ed7aca07227523fb12f35f937030bfa39e5f18d5134b2976e61bbb52703daa1aa5b816cec7081e1be04f2258962cf8954176ccfd27f677704edac643de73618af69e4407f53e71186f9e8f40440a3e6d5b80798565a0695f564439ae645c6c3cd62ec50debc78ca0cdcfee60f1b75fd3c1976340ceee7bec5d665cc316a2c0e4d554fc0da1fbd7c7189b6f2f6819faf7514fcc0104187c6e46079b59ce0bfe3fff27f2f37ba759232f1d3f3e13c12700281f82ab05eb0281a5ca586e7e70b607a74bdb2fc461a5081b3fbf8eb88f949c08a668ffc05a5e2b1b18969a86bea4e060e7c8fce9f8f3a16cb3f13bbbaeeb03dd094056cd06d2d20e7816ae5b6d2cb0f00d773321c5c951d046097c87c7fe20661e2adea4ad0c371abd78bbb8078211e26a1c216320c4ed1dac18a3cb6596091beb36c0a41215ccb9222ce1356f3d2e9844c797dcf569f386999701935332e03db2bc40ac6cdd6daabaaddc1c9f6c9de92bda356210c546c939ccfad9541927183048f4c667dd282c86b22c665bfc10d3fb55ccce81013d7b395fb937bf3f294e037b5514ba4b98fbd9be762e9b58bcab2a630aa21356fdfc692b9e1e904df1ff0bb7b9771dc02848b4cb4b573b0d60ffb4b33e696839f43af31ad5b67ae04ba602e0a3de3afb90bd622180bb94fe2bcaae57d7ccd62cb4b1216b5461f09bf57ed1306091c2105e14abae40dfedf5440075bc7823140e2a8516755a5f0622e5eb79a284a4ec623691d2f5700c5dca181d2f7bb85b1e0fec008a5a52fd4ca650b7e7cb4ad611925902afe448589513022fb96b86d234a0efabec78d1dc7676f8dfbe0d964364dc63a31f9daf1badc2fcf71393dacdc197fb1420bbdc0136c1be32805532e5516ac2f5951f8c93d99d56845eff0475e6e92df11582220893c6c04aa1b166d7a35918be405d9796b5982f843b4c425788882672689b05a52893ef1083b26be548b0b5b612834ffd254fc14f2618a9d9dc116753259a11973184d05b84b061bcf7fbc9f7fed3237b048ab9354b7e73fe237250f0a6aac0ce5f6058213196d0863b0eb3ef7fd39a14a6f4f7b97c8bc9409a72045e821fc782851e244ab89cf74d713db812b653c8b5d37b337c44cb21d3f1595c84d3d5b46c6d854bf7ad3345ca4b4a517960ebf359544753d4b50cde3869518c9e8778dcb854f06c4eb3075b3546c441b30f7c22a7d5b1079b069603a7af1fdf1b7d24384e2681397c380ff30a772e52f14af5302755072f09c4f7e2122de247671c91c4fb806141d2df61ca9ed029a806576151a0f4ca50baca810d813ef0c50a28bc35dbeddb8e0d0ad8ab550cf04c1fb89a9d143903e1b0597bc7f8f7057329e07c82f0364b973cc682c79dc65bcfe99c8ca57719c166451d3eaa789a67d8af3f0efe5fe65cc6308b36e063f158a7a3455e53c3fd134491aeb5141c4b424eb3cffedb095fdfb098b7f36b84fded2928b73024fa61a5f6079c7489ab3618082a58e16ced3f872e8061c0247227d18fd4fe75d9d8443e2567b0a576e7316679a9f2ea86243ad619fe4580f23d550acb6b5a069613c293ec1492cb796cd67576c275b2f40187804f8088c190842a49395e34acf7b376ad9924b9f6acdda4bed7993a4a0eb46e77f23c10f45a71f4229d316653804bd8ef625ec0b4cb8de035428857316811d85028ee1f1f9a34be84b03a47614313f77ae7b2a9284d4b92155ea4620e6bcd1f681cce98a983f45e1cc660646dc0409487d3c1de897d283cc45cf2cbe48569961dadb2f968d2b967b98f4291f3fc8b664126b9de8eb4f9012b737d3a5014d6ca8da30341a1fc9ebdbcc13d2d4bb2af5a8ca545893a9cf31e26149cd6bd64e2a4c5e1fd5d8a52de1bc251dbf5e655405dc7a7f69430773d5f68dce4d4621faab6f75ea4c79d7c68a56de1a4a52e356408217386b48c79f5d8658c4eec44ba5d5cdbe39dd6b71e713fdc90288faa23b096b7ee2fe61e79b1881d2592ca152cbb75f8e97b44904a84d22f601e2c43a40ef4f5000fcbdfe418554d486ab2403b67d1c56ae14a528b9785572fc1cc4b38ae6848eb4c803be2e59818f3b93d6b976e9d923086e56e8b8e2ebdb0bb586dd809429da4da305e2979cd5c3119146015953db65ee8b87c080376f147f5e12b5c2c13e03080964b33edf31ff9ea2268ca8cda1ba7778315da9a7b23db73296f8e11c2f7eb4c8bacfb3b229d88721bc836df07af186290124a5145a1504467cea78ae3e084705dfb665983625ccaede27c7d6f4b6eaba39374d1cfa9618b75ad236499e9136f9918c1c16fc2722fe9056e9fdf8d365d558cfdffbf75e181c9ab55db13e6fb1f0419ee2e2d71cb1cf5c220d6139e2979b268bd220b5c998b0305eec988a89e9337c379ccd1ffc0cf43918b6ea5d2e181a6812427f9b1d478d0e96d4e0f84c59e9697e5e4d3c46714b9';
+
+export {
+  otherWaysToGiveData,
+  tithingOfferingCards,
+  giveImpactData,
+  carouselCards,
+  moreGivingOpportunitiesId,
+  financialImprovementId,
+};
